requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[dotenv]
python-multipart
passlib[bcrypt]
email-validator
pandas
sentence-transformers
scikit-learn
mysql-connector-python
python-jose[cryptography]
python-dotenv
aiofiles
python-magic
PyPDF2
pytesseract
opencv-python
stripe


# Core Framework & Server
fastapi
uvicorn[standard]

# Database & ORM
sqlalchemy
asyncpg
alembic
databases

# Auth & Security
authlib
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
itsdangerous
email-validator

# File Uploads
aiofiles
python-multipart


# Optional: Email Services
aiohttp
httpx
aiosmtplib
jinja2



# Excel/CSV Parsing 
pandas
openpyxl
